+++ src/components/render/field/core.tsx
@@ -3,9 +3,7 @@ import { useState } from 'react';
 
 export const BasicFormField = ({ children, field }: BasicFieldProps) => (
   <FormField>
-    {field.type === 'hidden' ? null : (
-      <label className="label">{labelText(field)}</label>
-    )}
+    <Label field={field} />
     <FormControl>{children}</FormControl>
   </FormField>
 );
@@ -27,6 +25,11 @@ export const FormControl = (props: ParentProps) => (
 export const FormField = (props: ParentProps) => (
   <div className="field">{props.children}</div>
 );
+
+export const Label = ({ field }: FieldProp) =>
+  field.type === 'hidden' ? null : (
+    <label className="label">{labelText(field)}</label>
+  );
 
 export const labelText = (field: Field) => field.title ?? field.name;